Implemented photo strip.
Created thread to monitor creation of new threads and get them onto the UI in a semi-timely manner. The thread checks every 10 seconds. This sounds like a long time until you consider that new strips will be getting created at least 20ish seconds apart. Either way, time will tell if this is a bad plan or not.

Added new fields to the config to manage photo strip sizes and such.