initial commit of lua glue trigger module.
While it occurs to me that this might be largely pointless as Lua doesn't have real threads so any watcher will block, and as such a C module will need to be implemented and required by any lua module created, I can still hope that somebody might find this easier than creating an entire C module. 

If not, it only took a few hours to make this, and it was good practice and a good lesson learned, so w/e.