Major change to module_status_watchdog.
In the interest in reducing the number of threads used, decided to remove the thread for the module status watchdog.

In earlier versions of dmp photo booth, the module status watchdog thread used to call a now non-existent function on the modules: check status. Since there was no guarantee that this function would return immediately, it was not safe to add this functionality into the gtk main loop. Currently this function no longer exists, therefore it no longer makes sense to have an extra thread for the module status watchdog.

Removed the old watchdog thread function, the async queue, and the message handler. Added a new gtk idle function that replicates the functionality of the old watchdog thread/message handler pair.

Also removed finalize, since it no longer serves any purpose.